Fix count set
flag2 -> count, no longer need to downcast to byte

+++ NHSE.WinForms/Controls/ItemEditor.cs
@@ -34,7 +34,7 @@ namespace NHSE.WinForms
         {
             var id = WinFormsUtil.GetIndex(CB_ItemID);
             item.ItemId = (ushort) id;
-            item.Count = (byte) NUD_Count.Value;
+            item.Count = (ushort) NUD_Count.Value;
             item.UseCount = (ushort) NUD_Uses.Value;
             item.Flags0 = (byte) NUD_Flag0.Value;
             item.Flags1 = (byte) NUD_Flag1.Value;
